Lodha Kharadi Pune Residence
People with their growing needs tend to shift towards bigger homes and better spaces. The property buyers looking for homes with more rooms and better residential projects which fulfil their wish to have a grandeur lifestyle. Lodha Kharadi offers homes in 2 BHK with 3 BHK apartments. The project adds some of the most highlighted features of modern amenities and has a huge green landscape. It adds better residences with a price of Rs 1.0 Cr* lakhs only. The project has beautiful surroundings and making it one of the modern features to stay with tranquillity.

Lodha Kharadi Pune has come with some of the best apartments have one of the best locations with the Pune-Ahmednagar highway within proximity. It has spacious floors. The project is within a sprawling area of 14 acres. It comes with access to world-class security facilities. The project has a kid’s park for amusement, an indoor games club, a gym, a multipurpose hall and a fully equipped clubhouse. It has Jio fibre and Wi-Fi that gives faster internet connectivity and makes it easy for people who work from home. The project comes with uber-luxury living flats that are synonymous with giving a better lifestyle. The project gives living spaces with well-ventilated units. It has a spacious balcony that gives a beautiful view of the green central park.

Lodha Group has more than 4 decades of experience and comes with high-quality constructions. It is one of the most recognised real estate builders in the country. Lodha Altamount Homes is one of the premium homes in Pune-MMR and have been well planned that adds numerous benefits. One of the best is the location advantages which are close to all necessities within a few minutes away from it.

To know more - https://bit.ly/3zWJusq